+++ src/system/comercia/info.php
@@ -4,7 +4,13 @@ class Info
 {
     function IsInAdmin(){
         global $application_context;
-        return $application_context=="admin";
+        if (Util::version()->isMinimal("2.3")) {
+            return $application_context == "admin";
+        } else {
+            if (defined('DIR_CATALOG')) {
+                return $application_context = "admin";
+            }
+        }
     }
 }
 ?>